#780 Moved FunctionSpace into separate file, started some minor refactoring tests.
